+++ inlang/source-code/paraglide/paraglide-js-adapter-vite/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/paraglide-js-adapter-vite
 
+## 1.2.8
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.3.1
+
 ## 1.2.7
 
 ### Patch Changes